Read angle map values as size_t

+++ src/MolflowFacet.cpp
@@ -1351,7 +1351,7 @@ void InterfaceFacet::ImportAngleMap(const std::vector<std::vector<std::string>>&
 			for (size_t ix = 0; ix < phiWidth; ix++) {
 				size_t cellSize;
 				try {
-					angleMapCache[iy * phiWidth + ix] = std::stoi(table[iy + 1][ix + 1], &cellSize); //convert to double
+					angleMapCache[iy * phiWidth + ix] = std::stoull(table[iy + 1][ix + 1], &cellSize); //convert to size_t
 				}
 				catch (...) {
 					std::stringstream err;
